CounterTerm

This module provides the functions to manipulate counterterms.

ElectronLiquid.CounterTerm.chemicalpotential_renormalization — Method

function chemicalpotential_renormalization(order, data, δμ)

merge different diagrammatic orders with proper chemical potential renormalization

By definition, the chemical potential renormalization is defined as
Σ1 = Σ11
Σ2 = Σ20+Σ11*δμ1
Σ3 = Σ30+Σ11*δμ2+Σ12*δμ1^2+Σ21*δμ1
Σ4 = Σ40+Σ11*δμ3+Σ12*(2*δμ1*δμ2)+Σ13*δμ1^3+Σ21*δμ2+Σ22*δμ1^2+Σ31*δμ1

Arguments

order : total order data : Dict{OrderTuple, ActualData}, where OrderTuple is a tuple of two integer Tuple{NormalOrder+WOrder, GOrder} δμ : chemical potential renormalization for each order

ElectronLiquid.CounterTerm.getSigma — Method

function getSigma(para::ParaMC; order=para.order, parafile=parafileName)

read self-energy parameters from the file

Arguments

df : DataFrame paraid : Dictionary of parameter names and values order : the truncation order

ElectronLiquid.CounterTerm.mergeInteraction — Method

Merge interaction order and the main order (normalorder, Gorder, Worder) –> (normal+Worder, G_order)

ElectronLiquid.CounterTerm.renormalization — Function

function renormalization(order, data, δμ, δz=nothing; nbody=1, zrenorm=true)

First perform the chemical potential renormalization, then perform the z-factor renormalization

Arguments

order : total order data : Dict{OrderTuple, ActualData}, where OrderTuple is a tuple of two integer Tuple{NormalOrder+WOrder, GOrder} δμ : chemical potential counterterm δz : z-factor counterterm

zrenorm : turn on or off the z-factor renormalization nbody : nbody=1 for the one-body vertex function (self-energy, or Γ3) and nbody=2 for the two-body vertex function

ElectronLiquid.CounterTerm.sigmaCT — Function

function sigmaCT(order, μ, sw=Dict(key => 0.0 for key in keys(μ)); isfock=false)

Derive the chemicalpotential and z-factor counterterm for each order from the self-energy.

Arguments

order : total order μ : ReΣ(kF, w=0), Dict{OrderTuple, ActualData}, where OrderTuple is a tuple of two integer Tuple{NormalOrder+WOrder, GOrder}, or three integer Tuple{NormalOrder, WOrder, GOrder} sw : dImΣ(kF, w=0)/dw, Dict{OrderTuple, ActualData}, where OrderTuple is a tuple of two integer Tuple{NormalOrder+WOrder, GOrder}, or three integer Tuple{NormalOrder, WOrder, GOrder}

Return (δzi, δμ, δz)

The convention is the following: δzi : 1/z = 1+δzi1+δzi2+... δμ : chemicalpotentialshiftwithoutzrenormalization = δμ1+δμ2+... δz : z = 1+δz1+δz_2+...

Remark:

The chemical potential shift is the chemical potential shift without z-renormalization.

ElectronLiquid.CounterTerm.z_renormalization — Method

function z_renormalization(order, data, δz, nbody::Int)

By defintion, z=1+δz1+δz2+...

Then the z-factor renormalization resuffles the power series in the following way:
1. For the one-body vertex function
    DR = z*D ==> (D1+D2+D3+D4+...)*(1+δz1+δz2+δz3+...) = D1 + (D2+D1*δz1) + (D3+D2*δz1+D1*δz2) + (D4+D3*δz1+D2*δz2+D1*δz3) + ...
2. For the two-body vertex function
    DR = z^2*D ==> (D1+D2+D3+D4+...)*(1+δz1+δz2+δz3+...)^2 = D1 + (D2+2*D1*δz1) + (D3+2*D2*δz1+D1*(δz1^2+2*δz2) + (D4+2*D3*δz1+D2*(δz1^2+δz2)+2*D1*(δz1*δz2+δz3)) + ...

Note that the z-factor renormalization doesn't alter the W-order and the G-order, therefore, D_{m,n,k} renormalizes just as D_{m}.

Arguments

order : total order data : Vector of data, or Dict{OrderTuple, ActualData}, where OrderTuple is a tuple of two integer Tuple{NormalOrder+WOrder, GOrder} or three integer Tuple{NormalOrder, GOrder, W_Order} δz : z-factor renormalization for each order nbody : nbody=1 for the one-body vertex function (self-energy, or Γ3) and nbody=2 for the two-body vertex function