4-Vertex

This module provides the functions to calculate the 4-vertex function.

ElectronLiquid.Ver4.OneAngleAveraged — Type

struct OneAngleAveraged

The parameters for the one-angle-averaged vertex4.

Members

para: the parameters for the MC integration

kamp: the amplitude of the external momentum: [leftleg, rightlegs]

ωn: vector of the frequency of the external legs, each element is a 3-vector [leftin, leftout, right_in]

channel: the channel of the vertex4, :PH or :PP

l: the angular momentum of the angle average

ElectronLiquid.Ver4.PP_interaction — Function

function PP_interaction(para::ParaMC, kamp=para.kF, kamp2=para.kF; ct=true, kawargs...)

Calculate the interaction averged over the left incoming momentum (0, 0, kamp) and the left outgoing momentum kamp2*(cosθsinϕ, sinθsinϕ, cosϕ). The transfer momentum is q = sqrt(kamp^2 + kamp2^2 - 2 * cos(θ) * kamp * kamp2).

ElectronLiquid.Ver4._integrandGeneric — Method

Calculate vertex4 averged on the Fermi surface

ElectronLiquid.Ver4.exchange_interaction — Function

function exchange_interaction(para::ParaMC, kamp=para.kF, kamp2=para.kF; ct=true, kwargs...)

Calculate the exchange interaction between two particles with two incoming momentum kamp and kamp2. Averge over the angle between the two momenta.

ElectronLiquid.Ver4.integrandPH — Method

Calculate vertex4 averged on the Fermi surface

ElectronLiquid.Ver4.integrandPH_df — Method

Calculate vertex4 averged on the Fermi surface

ElectronLiquid.Ver4.one_angle_averaged — Method

function one_angle_averaged(paras::Vector{OneAngleAveraged}, diagram;
    neval=1e6, #number of evaluations
    print=0,
    alpha=3.0, #learning ratio
    config=nothing,
    kwargs...
)

Calculate the angle averaged vertex4 in the particle-hole or particle-particle channel.

Arguments

paras: vector of parameters for the one-angle-averaged vertex4

diagram: diagram to be calculated

neval: number of evaluations

print: print level

alpha: learning ratio

config: configuration for the integration